enabling uploads to be tested testing product file upload, #136

+++ app/helpers/amazonFileUploader.coffee
@@ -1,35 +1,51 @@
 AWS         = require 'aws-sdk'
 fs          = require 'fs'
-config      = require './config'
 path        = require 'path'
 os          = require 'os'
 ImageManipulation = require './imageManipulation'
 
 module.exports = class AmazonFileUploader
+  @configure = (id, secret, region, bucket) ->
+    unless @dryrun
+      AWS.config.update accessKeyId: id, secretAccessKey: secret, region: region
+      @bucket = bucket
+      @im = new ImageManipulation()
+
+  @classProperty 'dryrun',
+    get: =>
+      @_dryrun = off unless @_dryrun?
+      @_dryrun
+    set: (val) =>
+      @_dryrun = val
+      if val
+        @bucket = 'dryrun'
+        @filesUploaded = []
+
   constructor: ->
-    AWS.config.update accessKeyId: config.aws.accessKeyId, secretAccessKey: config.aws.secretKey, region: config.aws.region
     @s3 = new AWS.S3()
-    @bucket = config.aws.imagesBucket
-    @tmpdir = os.tmpdir()
-    @im = new ImageManipulation()
   upload: (fileName, file, dimensions, cb) ->
-    @im.isSizeSmallerThan file.path, dimensions, (err, itIs) =>
+    if AmazonFileUploader.dryrun
+      console.log "NOT uploading file, dry run"
+      url = "https://s3.amazonaws.com/#{AmazonFileUploader.bucket}/#{fileName}"
+      AmazonFileUploader.filesUploaded.push url
+      return cb null, url
+    AmazonFileUploader.im.isSizeSmallerThan file.path, dimensions, (err, itIs) =>
       return cb err if err?
       return cb smallerThan: "O tamanho da imagem é menor do que o esperado, ela deve ter no mínimo #{dimensions}." if itIs
-      @im.resizeAndCrop file.path, dimensions, (err, newFileName) =>
+      AmazonFileUploader.im.resizeAndCrop file.path, dimensions, (err, newFileName) =>
         return cb err if err?
-        fileParams = Bucket: @bucket, Key: fileName, Body: fs.createReadStream(newFileName), ContentType: file.headers['content-type']
+        fileParams = Bucket: AmazonFileUploader.bucket, Key: fileName, Body: fs.createReadStream(newFileName), ContentType: file.headers['content-type']
         @s3.putObject fileParams, (err, data) =>
           return cb err if err?
-          url = "https://s3.amazonaws.com/#{@bucket}/#{fileName}"
+          url = "https://s3.amazonaws.com/#{AmazonFileUploader.bucket}/#{fileName}"
           cb null, url
           fs.unlink newFileName
   delete: (fileName, cb) ->
-    key = fileName.replace "https://s3.amazonaws.com/#{@bucket}/", ""
-    fileParams = Bucket: @bucket, Key: key
+    key = fileName.replace "https://s3.amazonaws.com/#{AmazonFileUploader.bucket}/", ""
+    fileParams = Bucket: AmazonFileUploader.bucket, Key: key
     @s3.deleteObject fileParams, (err, data) => cb err
   getFileNameFromFullName: (fileName) ->
-    key = fileName.replace "https://s3.amazonaws.com/#{@bucket}/", ""
+    key = fileName.replace "https://s3.amazonaws.com/#{AmazonFileUploader.bucket}/", ""
     key
   randomName: (folder, file) ->
     rand = -> Math.random() * Math.pow(10, 18)
